topコンポーネントを分割しResultコンポーネントに一覧表示処理を記載
