+++ frontend/app/pages/Top.tsx
@@ -1,4 +1,6 @@
 import { useState } from "react";
+import Result from "./components/Result";
+
 type Muscle = {
   id: number;
   name: string;
@@ -21,31 +23,7 @@ export function Top() {
     <div className="top">
       <h1>筋トレ実績</h1>
       <button onClick={getMuscle}>一覧取得</button>
-      <div>
-        <table>
-          <thead>
-            <tr>
-              <th className="muscle-header">種目名</th>
-              <th className="muscle-header">実施日</th>
-              <th className="muscle-header">回数</th>
-            </tr>
-          </thead>
-          <tbody>
-            {muscle.map((c) => {
-              const date = new Date(c.date);
-              const formattedDate = date.toLocaleDateString("ja-JP");
-
-              return (
-                <tr key={c.id}>
-                  <th className="muscle-cell">{c.name}</th>
-                  <th className="muscle-cell">{formattedDate}</th>
-                  <th className="muscle-cell">{c.count}</th>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
+      <Result muscle={muscle} />
     </div>
   );
 }
